backend/requirements.txt
django
djangorestframework
django-cors-headers
opencv-python
numpy
scikit-learn
pandas
Pillow
